rejig load_to_target step to use shared bits
