+++ migration_steps/acquire_target_ids/app/entities/client.py
@@ -1,5 +1,5 @@
 import os
-from db_helpers import *
+from db_helpers import result_from_sql_file, df_from_sql_file, execute_insert, execute_sql_file
 from pathlib import Path
 
 current_path = Path(os.path.dirname(os.path.realpath(__file__)))
