+++ migration_steps/load_to_target-dockerfile
@@ -1,7 +1,8 @@
 FROM opg_casrec_migration_base_image:latest
-COPY prepare_target/requirements.txt /migration/requirements.txt
+COPY load_to_target/requirements.txt /migration/requirements.txt
 RUN pip3 install -r /migration/requirements.txt
-COPY prepare_target/app /migration/app
+COPY load_to_target/app /migration/app
 COPY shared /shared
 WORKDIR /migration/app
+RUN chmod 755 /migration/app/*.py
 CMD ["echo", "NO_OP"]
