--- migration_steps/load_to_target/Dockerfile
@@ -1,7 +1,0 @@
-FROM opg_casrec_migration_base_image:latest
-COPY ./requirements.txt /requirements.txt
-RUN pip3 install -r requirements.txt
-COPY app /app
-WORKDIR app
-RUN chmod 755 *.py
-CMD ["echo", "NO_OP"]
