+++ migration_steps/load_to_target/app/entities/address.py
@@ -1,9 +1,14 @@
-from helpers import *
+import os
+from db_helpers import df_from_sql_file, execute_update, execute_insert, execute_sql_file
+from pathlib import Path
+
+current_path = Path(os.path.dirname(os.path.realpath(__file__)))
+sql_path = current_path / '../sql'
 
 
 def target_update(config, conn_migration, conn_target):
     schema = config.schemas['pre_migrate']
-    addresses_df = df_from_sql_file('get_skeleton_addresses.sql', conn_migration, schema)
+    addresses_df = df_from_sql_file(sql_path, 'get_skeleton_addresses.sql', conn_migration, schema)
 
     # transpose id column
     addresses_df = addresses_df.drop(["id", "sirius_id", "person_id"], axis=1)
@@ -21,7 +26,7 @@ def target_update(config, conn_migration, conn_target):
 
 def target_add(config, conn_migration, conn_target):
     schema = config.schemas['pre_migrate']
-    addresses_df = df_from_sql_file('get_new_addresses.sql', conn_migration, schema)
+    addresses_df = df_from_sql_file(sql_path, 'get_new_addresses.sql', conn_migration, schema)
 
     # don't send id
     addresses_df = addresses_df.drop(["id", "sirius_id"], axis=1)
@@ -41,4 +46,4 @@ def target_add(config, conn_migration, conn_target):
 
 def reindex_target_ids(config, conn_migration):
     schema = config.schemas['pre_migrate']
-    execute_sql_file('merge_target_addresses.sql', conn_migration, schema)
+    execute_sql_file(sql_path, 'merge_target_addresses.sql', conn_migration, schema)
