+++ migration_steps/load_to_target/app/entities/client.py
@@ -1,9 +1,14 @@
-from helpers import *
+import os
+from db_helpers import df_from_sql_file, execute_update, result_from_sql_file, execute_insert
+from pathlib import Path
+
+current_path = Path(os.path.dirname(os.path.realpath(__file__)))
+sql_path = current_path / '../sql'
 
 
 def target_update(config, conn_migration, conn_target):
     schema = config.schemas['pre_migrate']
-    persons_df = df_from_sql_file('get_skeleton_clients.sql', conn_migration, schema)
+    persons_df = df_from_sql_file(sql_path, 'get_skeleton_clients.sql', conn_migration, schema)
 
     # transpose id column
     persons_df = persons_df.drop(["id", "sirius_id"], axis=1)
@@ -17,7 +22,7 @@ def target_update(config, conn_migration, conn_target):
 
 def target_add(config, conn_migration, conn_target):
     schema = config.schemas['pre_migrate']
-    persons_df = df_from_sql_file('get_new_clients.sql', conn_migration, schema)
+    persons_df = df_from_sql_file(sql_path, 'get_new_clients.sql', conn_migration, schema)
 
     # don't send id
     persons_df = persons_df.drop(["id", "sirius_id"], axis=1)
@@ -28,7 +33,7 @@ def target_add(config, conn_migration, conn_target):
 
     # uid not implemented upstream so here's a workaround
     rowcount = len(persons_df.index)
-    max_person_uid = result_from_sql_file('get_max_person_uid.sql', conn_target)
+    max_person_uid = result_from_sql_file(sql_path, 'get_max_person_uid.sql', conn_target)
     persons_df["uid"] = list(range(max_person_uid + 1, max_person_uid + rowcount + 1, 1))
 
     execute_insert(conn_target, persons_df, 'persons')
@@ -36,7 +41,7 @@ def target_add(config, conn_migration, conn_target):
 
 def reindex_target_ids(config, conn_migration, conn_target):
     schema = config.schemas['pre_migrate']
-    sirius_persons_df = df_from_sql_file('select_sirius_clients.sql', conn_target)
+    sirius_persons_df = df_from_sql_file(sql_path, 'select_sirius_clients.sql', conn_target)
 
     cursor = conn_migration.cursor()
     cursor.execute('TRUNCATE pre_migrate.sirius_map_clients;')
